+++ httpapi_exporter.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"net/http"
 	"os"
+	"regexp"
 	"time"
 
 	_ "net/http/pprof"
@@ -17,6 +18,8 @@ import (
 	"github.com/prometheus/common/promlog"
 	"github.com/prometheus/common/promlog/flag"
 	"github.com/prometheus/common/version"
+	"github.com/prometheus/exporter-toolkit/web"
+	"github.com/prometheus/exporter-toolkit/web/kingpinflag"
 )
 
 const (
@@ -26,18 +29,88 @@ const (
 )
 
 var (
-	listenAddress = kingpin.Flag("web.listen-address", "The address to listen on for HTTP requests.").Default(metricsPublishingPort).String()
-	metricsPath   = kingpin.Flag("web.telemetry-path", "Path under which to expose collector's internal metrics.").Default("/metrics").String()
-	configFile    = kingpin.Flag("config.file", "Exporter configuration file.").Short('c').Default("config/config.yml").String()
+	// listenAddress = kingpin.Flag("web.listen-address", "The address to listen on for HTTP requests.").Default(metricsPublishingPort).String()
+	metricsPath = kingpin.Flag("web.telemetry-path", "Path under which to expose collector's internal metrics.").Default("/metrics").String()
+	configFile  = kingpin.Flag("config.file", "Exporter configuration file.").Short('c').Default("config/config.yml").String()
 	//debug_flag    = kingpin.Flag("debug", "debug connection checks.").Short('d').Default("false").Bool()
 	dry_run = kingpin.Flag("dry-run", "Only check exporter configuration file and exit.").Short('n').Default("false").Bool()
 	// alsologtostderr = kingpin.Flag("alsologtostderr", "log to standard error as well as files.").Default("true").Bool()
 	target_name    = kingpin.Flag("target", "In dry-run mode specify the target name, else ignored.").Short('t').String()
 	collector_name = kingpin.Flag("metric", "Specify the collector name restriction to collect, replace the collector_names set for each target.").Short('m').String()
+	toolkitFlags   = kingpinflag.AddFlags(kingpin.CommandLine, metricsPublishingPort)
 )
 
 func init() {
 	prometheus.MustRegister(version.NewCollector(exporter_name))
+}
+
+const (
+	OpEgals = 1
+	OpMatch = 2
+)
+
+type route struct {
+	path    string
+	regex   *regexp.Regexp
+	handler http.HandlerFunc
+}
+
+func newRoute(op int, path string, handler http.HandlerFunc) *route {
+	if op == OpEgals {
+		return &route{path, nil, handler}
+	} else if op == OpMatch {
+		return &route{"", regexp.MustCompile("^" + path + "$"), handler}
+
+	} else {
+		return nil
+	}
+
+}
+func BuildHandler(exporter Exporter) http.Handler {
+	var routes = []*route{
+		newRoute(OpEgals, "/healthz", func(w http.ResponseWriter, r *http.Request) { http.Error(w, "OK", http.StatusOK) }),
+		newRoute(OpEgals, "/", HomeHandlerFunc(*metricsPath, exporter)),
+		newRoute(OpEgals, "/status", StatusHandlerFunc(*metricsPath, exporter)),
+		newRoute(OpEgals, "/config", ConfigHandlerFunc(*metricsPath, exporter)),
+		newRoute(OpEgals, "/targets", TargetsHandlerFunc(*metricsPath, exporter)),
+		newRoute(OpEgals, *metricsPath, func(w http.ResponseWriter, r *http.Request) { ExporterHandlerFor(exporter).ServeHTTP(w, r) }),
+		// Expose exporter metrics separately, for debugging purposes.
+		newRoute(OpEgals, "/httpapi_exporter_metrics", func(w http.ResponseWriter, r *http.Request) { promhttp.Handler().ServeHTTP(w, r) }),
+
+		// pprof handle
+		newRoute(OpMatch, "/debug/.+", http.DefaultServeMux.ServeHTTP),
+		// newRoute("/debug/pprof/cmdline", pprof.Cmdline),
+		// newRoute("/debug/pprof/profile", pprof.Profile),
+		// newRoute("/debug/pprof/symbol", pprof.Symbol),
+		// newRoute("/debug/pprof/trace", pprof.Trace),
+		// newRoute("default", func(w http.ResponseWriter, r *http.Request) { http.Handler() .ServeHTTP(w, r) }),
+	}
+
+	return http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		for _, route := range routes {
+			if route == nil {
+				continue
+			}
+			if route.regex != nil {
+				if route.regex.MatchString(req.URL.Path) {
+					route.handler(w, req)
+					return
+				}
+			} else if req.URL.Path == route.path {
+				route.handler(w, req)
+				return
+			}
+		}
+		// http.DefaultServeMux.ServeHTTP(w, req)
+		// h, _ := http.DefaultServeMux.Handler(req)
+		// if h != nil {
+		// 			h.ServeHTTP(w, req)
+		// } else {
+		err := fmt.Errorf("not found")
+		HandleError(http.StatusNotFound, err, *metricsPath, exporter, w, req)
+		// }
+	})
+
 }
 
 func main() {
@@ -119,19 +192,29 @@ func main() {
 		os.Exit(0)
 	}
 
-	// Setup and start webserver.
-	http.HandleFunc("/healthz", func(w http.ResponseWriter, r *http.Request) { http.Error(w, "OK", http.StatusOK) })
-	http.HandleFunc("/", HomeHandlerFunc(*metricsPath, exporter))
-	http.HandleFunc("/config", ConfigHandlerFunc(*metricsPath, exporter))
-	http.HandleFunc("/status", StatusHandlerFunc(*metricsPath, exporter))
-	http.HandleFunc("/targets", TargetsHandlerFunc(*metricsPath, exporter))
-	http.Handle(*metricsPath, ExporterHandlerFor(exporter))
-	// Expose exporter metrics separately, for debugging purposes.
-	http.Handle("/httpapi_exporter_metrics", promhttp.Handler())
+	exporter.SetStartTime(time.Now())
 
-	level.Info(logger).Log("msg", "Listening on address", "address", *listenAddress)
-	if err := http.ListenAndServe(*listenAddress, nil); err != nil {
-		level.Error(logger).Log("msg", "Error starting HTTP server", "errmsg", err)
+	// Setup and start webserver.
+	server := &http.Server{
+		Handler: BuildHandler(exporter),
+	}
+	if err := web.ListenAndServe(server, toolkitFlags, logger); err != nil {
+		level.Error(logger).Log("err", err)
 		os.Exit(1)
 	}
+
+	// http.HandleFunc("/healthz", func(w http.ResponseWriter, r *http.Request) { http.Error(w, "OK", http.StatusOK) })
+	// http.HandleFunc("/", HomeHandlerFunc(*metricsPath, exporter))
+	// http.HandleFunc("/config", ConfigHandlerFunc(*metricsPath, exporter))
+	// http.HandleFunc("/status", StatusHandlerFunc(*metricsPath, exporter))
+	// http.HandleFunc("/targets", TargetsHandlerFunc(*metricsPath, exporter))
+	// http.Handle(*metricsPath, ExporterHandlerFor(exporter))
+	// // Expose exporter metrics separately, for debugging purposes.
+	// http.Handle("/httpapi_exporter_metrics", promhttp.Handler())
+
+	// level.Info(logger).Log("msg", "Listening on address", "address", *listenAddress)
+	// if err := http.ListenAndServe(*listenAddress, nil); err != nil {
+	// 	level.Error(logger).Log("msg", "Error starting HTTP server", "errmsg", err)
+	// 	os.Exit(1)
+	// }
 }