+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a changelog](https://github.com/olivierlacan/keep-a-changelog).
 
  <!--next-version-placeholder-->
+## 0.3.1 / 2023-09-24
+- use standard prometheus args web.listen-address and web.config.file for https
+- add server start_time in /status page
+- modify http server routing process
+
 ## 0.3.0 / 2023-09-24
 ### Changed
 - Initial release